making it neater
