+++ src/prompt.c
@@ -6,40 +6,11 @@
 /*   By: seayeo <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/07 14:04:27 by seayeo            #+#    #+#             */
-/*   Updated: 2024/06/20 16:56:44 by seayeo           ###   ########.fr       */
+/*   Updated: 2024/06/23 12:51:23 by seayeo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-char	*findpath(char *envp[])
-{
-	int	i;
-
-	i = 0;
-	while (envp[i])
-	{
-		// printf("%s\n", envp[i]);
-		if (ft_strncmp(envp[i], "PATH=", 5) == 0)
-			return (envp[i]);
-		i++;
-	}
-	return (NULL);
-}
-
-char	*finduser(char *envp[])
-{
-	int	i;
-
-	i = 0;
-	while (envp[i])
-	{
-		if (ft_strncmp(envp[i], "USER=", 5) == 0)
-			return(ft_substr(envp[i], 5, 6));
-		i++;
-	}
-	return (NULL);
-}
 
 char	*findhost(char *envp[])
 {
@@ -65,7 +36,6 @@ char	*form_prompt(char *cwd)
 	char	*temp;
 	char	*ret;
 
-	// username = finduser(envp);
 	username = getenv("USER");
 	// host = getenv("session_manager");
 	// temp = ft_strjoin(username, "@");
